chore: Add examples directory
- Moved `synopsis.rs` to examples dir
- Added `examples/common.rs` to contain the example wallet
- Update README to link to the examples
